Factor out KiezCard
