+++ src/pages/Spielstrassen/pages/Kieze.tsx
@@ -1,18 +1,14 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React from 'react';
-import { generatePath } from 'react-router-dom';
-import { Paper, Container } from '@material-ui/core';
+import { Container } from '@material-ui/core';
 import styled from 'styled-components';
-import slugify from 'slugify';
 
 import Button from '~/components2/Button';
-import Link from '~/components/Link';
 import { Insert as ImageInsert } from '~/components2/Image';
 import KiezKarte from '~/images/spielstrassen/kiezkarte@3x.jpg';
-import Supporter from '~/images/spielstrassen/supporter.svg';
-import SupporterCheck from '~/images/spielstrassen/supporter-check.svg';
 import config from '~/pages/Spielstrassen/config';
 import Header from '../components/Header';
+import KiezCard from '../components/KiezCard';
 import { media } from '~/styles/utils';
 
 const kiezData = [
@@ -48,41 +44,6 @@ const KiezListing = styled.div`
   margin: 1em 0 2em;
 `;
 
-const KiezPaper = styled(Paper)`
-  padding: 0.5em;
-  margin-bottom: 1em;
-
-  dl {
-    margin-top: 0;
-  }
-
-  dd,
-  dt {
-    display: inline;
-  }
-
-  dd {
-    margin-left: 0.5em;
-  }
-
-  footer {
-    display: flex;
-
-    .supportercount {
-      font-size: 0.75em;
-      line-height: 1.5em;
-      color: ${config.colors.darkgrey};
-      hyphens: manual;
-      margin-left: 1em;
-      margin-right: 0.5em;
-    }
-
-    button {
-      margin-left: auto;
-    }
-  }
-`;
-
 const ContactButton = styled(Button)`
   margin-bottom: 2em;
   width: 100%;
@@ -92,12 +53,6 @@ const ContactButton = styled(Button)`
     right: 24px;
     width: initial;
   `}
-`;
-
-const SupportersReached = styled(SupporterCheck)`
-  margin-left: -14px;
-  margin-top: -5px;
-  visibility: ${(props) => (props.visible ? 'visible' : 'hidden')};
 `;
 
 const ZoomMap = styled.span`
@@ -124,38 +79,6 @@ const ZoomMap = styled.span`
   }
 `;
 
-const Kiez = ({ name, street, supporters = 0 }) => {
-  const signupUrl = generatePath(config.routes.spielstrassen.register, {
-    kiez: slugify(name, { lower: true })
-  });
-
-  return (
-    <KiezPaper elevation={5}>
-      <dl>
-        <strong>
-          <dt>Spielstraße:</dt>
-          <dd>{street}</dd>
-        </strong>
-        <br />
-        <dt>Kiez:</dt>
-        <dd>{name}</dd>
-      </dl>
-      <footer>
-        <Supporter />
-        <SupportersReached
-          visible={supporters >= config.spielstrassen.supporterGoal}
-        />
-        <span className="supportercount">
-          {supporters} Unter&shy;stützer registriert
-        </span>
-        <Button flat>
-          <Link to={signupUrl}>Unterstützen</Link>
-        </Button>
-      </footer>
-    </KiezPaper>
-  );
-};
-
 const Kieze = () => (
   <>
     <Header showInfoLink />
@@ -169,7 +92,7 @@ const Kieze = () => (
       </ZoomMap>
       <KiezListing>
         {kiezData.map(({ name, supporters, street }) => (
-          <Kiez street={street} name={name} supporters={supporters} />
+          <KiezCard street={street} name={name} supporters={supporters} />
         ))}
       </KiezListing>
       <p>
